OL-72_if-check-to-make-sure-picture-exists-first

+++ app/Http/Controllers/AuthorController.php
@@ -91,7 +91,9 @@
          */
         public function destroy(Author $author)
         {
-            Storage::disk('public')->delete($author->picture);
+            if (Storage::disk('public')->exists($author->picture)) {
+                Storage::disk('public')->delete($author->picture);
+            }
             $author->delete();
 
             return response()->json(['message' => 'Author Deleted Successfully']);
